requirements.txt
PySide6==6.6.1
psycopg2-binary==2.9.7